Code Generation

Types

GraphComputing.Tape — Type

Tape{INPUT}

TODO: update docs

INPUT the input type of the problem instance

code::Vector{Expr}: The julia expression containing the code for the whole graph.

inputSymbols::Dict{String, Vector{Symbol}}: A dictionary of symbols mapping the names of the input nodes of the graph to the symbols their inputs should be provided on.

outputSymbol::Symbol: The symbol of the final calculated value

Function Generation

Implementations for generation of a callable function. A function generated this way cannot immediately be called. One Julia World Age has to pass before this is possible, which happens when the global Julia scope advances. If the DAG and therefore the generated function becomes too large, use the tape machine instead, since compiling large functions becomes infeasible.

GraphComputing.execute — Method

execute(
    graph::DAG,
    instance,
    machine::Machine,
    input,
    context_module::Module
)

Execute the code of the given graph on the given input values.

This is essentially shorthand for

tape = gen_tape(graph, instance, machine, context_module)
return execute_tape(tape, input)

GraphComputing.get_compute_function — Method

get_compute_function(
    graph::DAG,
    instance,
    machine::Machine,
    context_module::Module
)

Return a function of signature compute_<id>(input::input_type(instance)), which will return the result of the DAG computation on the given input. The final argument context_module should always be @__MODULE__ to be able to use functions defined in the caller's environment. For this to work, you need

using RuntimeGeneratedFunctions
RuntimeGeneratedFunctions.init(@__MODULE__)

in your top level.

GraphComputing.get_cuda_kernel — Method

get_cuda_kernel(
    graph::DAG,
    instance,
    machine::Machine,
)

Return a function of signature compute_<id>(input::CuVector, output::CuVector, n::Int64), which will return the result of the DAG computation of the input on the given output variable.

Tape Machine

GraphComputing.call_fc — Method

call_fc(fc::FunctionCall, cache::Dict{Symbol, Any})

Execute the given FunctionCall on the dictionary.

Several more specialized versions of this function exist to reduce vector unrolling work for common cases.

GraphComputing.execute_tape — Method

execute_tape(tape::Tape, input::Input) where {Input}

Execute the given tape with the given input.

For implementation reasons, this disregards the set CacheStrategy of the devices and always uses a dictionary.

GraphComputing.expr_from_fc — Method

expr_from_fc(fc::FunctionCall)

For a given function call, return an expression evaluating it.

GraphComputing.gen_cache_init_code — Method

gen_cache_init_code(machine::Machine)

For each AbstractDevice in the given Machine, returning a Vector{Expr} doing the initialization.

GraphComputing.gen_input_assignment_code — Method

gen_input_assignment_code(
    input_symbols::Dict{String, Vector{Symbol}},
    instance::AbstractProblemInstance,
    machine::Machine,
    context_module::Module
)

Return a Vector{Expr} doing the input assignments from the given problem_input onto the input_symbols.

GraphComputing.gen_tape — Function

gen_tape(
    graph::DAG,
    instance::AbstractProblemInstance,
    machine::Machine,
    context_module::Module,
    scheduler::AbstractScheduler = GreedyScheduler()
)

Generate the code for a given graph. The return value is a Tape.

See also: execute, execute_tape