Graph

Type

GraphComputing.DAG — Type

DAG

The representation of the graph as a set of Nodes.

A DAG can be loaded using the appropriate parsedag function, e.g. [`parsedag`](@ref).

Operations can be applied on it using push_operation! and reverted using pop_operation! like a stack. To get the set of possible operations, use get_operations. The members of the object should not be manually accessed, instead always use the provided interface functions.

GraphComputing.DAG — Method

DAG()

Construct and return an empty DAG.

GraphComputing.PossibleOperations — Type

PossibleOperations

A struct storing all possible operations on a DAG. To get the PossibleOperations on a DAG, use get_operations.

GraphComputing.PossibleOperations — Method

PossibleOperations()

Construct and return an empty PossibleOperations object.

Interface

GraphComputing.can_pop — Method

can_pop(graph::DAG)

Return true if pop_operation! is possible, false otherwise.

GraphComputing.pop_operation! — Method

pop_operation!(graph::DAG)

Revert the latest applied operation on the graph.

See also: DAG, push_operation!

GraphComputing.push_operation! — Method

push_operation!(graph::DAG, operation::Operation)

Apply a new operation to the graph.

See also: DAG, pop_operation!

GraphComputing.reset_graph! — Method

reset_graph!(graph::DAG)

Reset the graph to its initial state with no operations applied.

Compare

Base.in — Method

in(edge::Edge, graph::DAG)

Check whether the edge is part of the graph.

Base.in — Method

in(node::Node, graph::DAG)

Check whether the node is part of the graph.

Mute

GraphComputing._insert_edge! — Method

_insert_edge!(graph::DAG, node1::Node, node2::Node; track = true, invalidate_cache = true)

Insert the edge between node1 (child) and node2 (parent) into the graph.

Warning

For creating new graphs, use the public version insert_edge! instead which uses the defaults false for the keywords.

Keyword Arguments

track::Bool: Whether to add the changes to the DAG's Diff. Should be set false in parsing or graph creation functions for performance.

invalidate_cache::Bool: Whether to invalidate caches associated with the changes. Should also be turned off for graph creation or parsing.

See also: _insert_node!, _remove_node!, _remove_edge!

GraphComputing._insert_node! — Method

_insert_node!(graph::DAG, node::Node; track = true, invalidate_cache = true)

Insert the node into the graph.

Warning

For creating new graphs, use the public version insert_node! instead which uses the defaults false for the keywords.

Keyword Arguments

track::Bool: Whether to add the changes to the DAG's Diff. Should be set false in parsing or graph creation functions for performance.

invalidate_cache::Bool: Whether to invalidate caches associated with the changes. Should also be turned off for graph creation or parsing.

See also: _remove_node!, _insert_edge!, _remove_edge!

GraphComputing._remove_edge! — Method

_remove_edge!(graph::DAG, node1::Node, node2::Node; track = true, invalidate_cache = true)

Remove the edge between node1 (child) and node2 (parent) into the graph.

Keyword Arguments

track::Bool: Whether to add the changes to the DAG's Diff. Should be set false in parsing or graph creation functions for performance.

invalidate_cache::Bool: Whether to invalidate caches associated with the changes. Should also be turned off for graph creation or parsing.

See also: _insert_node!, _remove_node!, _insert_edge!

GraphComputing._remove_node! — Method

_remove_node!(graph::DAG, node::Node; track = true, invalidate_cache = true)

Remove the node from the graph.

Keyword Arguments

track::Bool: Whether to add the changes to the DAG's Diff. Should be set false in parsing or graph creation functions for performance.

invalidate_cache::Bool: Whether to invalidate caches associated with the changes. Should also be turned off for graph creation or parsing.

See also: _insert_node!, _insert_edge!, _remove_edge!

GraphComputing.get_snapshot_diff — Method

get_snapshot_diff(graph::DAG)

Return the graph's Diff since last time this function was called.

See also: revert_diff!, AppliedOperation and revert_operation!

GraphComputing.insert_edge! — Method

insert_edge!(graph::DAG, node1::Node, node2::Node)

Insert the edge between node1 (child) and node2 (parent) into the graph.

GraphComputing.insert_node! — Method

insert_node!(graph::DAG, node::Node)

Insert the node into the graph.

GraphComputing.invalidate_caches! — Method

invalidate_caches!(graph::DAG, operation::NodeReduction)

Invalidate the operation caches for a given NodeReduction.

This deletes the operation from the graph's possible operations and from the involved nodes' own operation caches.

GraphComputing.invalidate_caches! — Method

invalidate_caches!(graph::DAG, operation::NodeSplit)

Invalidate the operation caches for a given NodeSplit.

This deletes the operation from the graph's possible operations and from the involved nodes' own operation caches.

GraphComputing.invalidate_operation_caches! — Method

invalidate_operation_caches!(graph::DAG, node::ComputeTaskNode)

Invalidate the operation caches of the given node through calls to the respective invalidate_caches! functions.

GraphComputing.invalidate_operation_caches! — Method

invalidate_operation_caches!(graph::DAG, node::DataTaskNode)

Invalidate the operation caches of the given node through calls to the respective invalidate_caches! functions.

Print

Base.show — Method

show(io::IO, graph::DAG)

Print the given graph to io. If there are too many nodes it will print only a summary of them.

GraphComputing.show_nodes — Method

show_nodes(io::IO, graph::DAG)

Print a graph's nodes. Should only be used for small graphs as it prints every node in a list.

Properties

GraphComputing.get_entry_nodes — Method

get_entry_nodes(graph::DAG)

Return a vector of the graph's entry nodes.

GraphComputing.get_exit_node — Method

get_exit_node(graph::DAG)

Return the graph's exit node. This assumes the graph only has a single exit node. If the graph has multiple exit nodes, the one encountered first will be returned.

GraphComputing.get_properties — Method

get_properties(graph::DAG)

Return the graph's GraphProperties.

GraphComputing.operation_stack_length — Method

operation_stack_length(graph::DAG)

Return the number of operations applied to the graph.

Validate

GraphComputing.is_connected — Method

is_connected(graph::DAG)

Return whether the given graph is connected.

GraphComputing.is_scheduled — Method

is_scheduled(graph::DAG)

Validate that the entire graph has been scheduled, i.e., every ComputeTaskNode has its .device set.

GraphComputing.is_valid — Method

is_valid(graph::DAG)

Validate the entire graph using asserts. Intended for testing with @assert is_valid(graph).