Task

Type

GraphComputing.AbstractComputeTask — Type

AbstractComputeTask <: AbstractTask

The shared base type for any compute task.

GraphComputing.AbstractDataTask — Type

AbstractDataTask <: AbstractTask

The shared base type for any data task.

GraphComputing.AbstractTask — Type

AbstractTask

The shared base type for any task.

GraphComputing.DataTask — Type

DataTask <: AbstractDataTask

Task representing a specific data transfer.

Create

Base.copy — Method

copy(t::AbstractComputeTask)

Return a copy of the given compute task.

Base.copy — Method

copy(t::AbstractDataTask)

Fallback implementation of the copy of an abstract data task, throwing an error.

Compare

Base.:== — Method

==(t1::AbstractComputeTask, t2::AbstractComputeTask)

Equality comparison between two compute tasks.

Base.:== — Method

==(t1::AbstractDataTask, t2::AbstractDataTask)

Equality comparison between two data tasks.

Base.:== — Method

==(t1::AbstractTask, t2::AbstractTask)

Fallback implementation of equality comparison between two abstract tasks. Always returns false. For equal specific types of t1 and t2, a more specific comparison is called instead, doing an actual comparison.

Compute

GraphComputing.get_function_call — Method

get_function_call(n::Node)
get_function_call(t::AbstractTask, device::AbstractDevice, in_symbols::AbstractVector, out_symbol::Symbol)

For a node or a task together with necessary information, return a vector of FunctionCalls for the computation of the node or task.

For ordinary compute or data tasks the vector will contain exactly one element.

Properties

Base.copy — Method

copy(t::DataTask)

Copy the data task and return it.

GraphComputing.children — Method

children(::DataTask)

Return the number of children of a data task (always 1).

GraphComputing.compute — Function

compute(t::AbstractTask; data...)

Fallback implementation of the compute function of a compute task, throwing an error.

GraphComputing.compute_effort — Function

compute_effort(t::AbstractTask)

Fallback implementation of the compute effort of a task, throwing an error.

GraphComputing.compute_effort — Method

compute_effort(t::AbstractDataTask)

Return the compute effort of a data task, always zero, regardless of the specific task.

GraphComputing.data — Function

data(t::AbstractTask)

Fallback implementation of the data of a task, throwing an error.

GraphComputing.data — Method

data(t::AbstractComputeTask)

Return the data of a compute task, always zero, regardless of the specific task.

GraphComputing.data — Method

data(t::AbstractDataTask)

Return the data of a data task. Given by the task's .data field.