Devices

Interface

GraphComputing.AbstractDevice — Type

AbstractDevice

Abstract base type for every device, like GPUs, CPUs or any other compute devices. Every implementation needs to implement various functions and needs a member cacheStrategy.

GraphComputing.CacheStrategy — Type

CacheStrategy

Abstract base type for caching strategies.

See also: strategies

GraphComputing.Dictionary — Type

Dictionary <: CacheStrategy

A caching strategy relying on a dictionary of Symbols to store every input and output.

Implements the CacheStrategy interface.

GraphComputing.LocalVariables — Type

LocalVariables <: CacheStrategy

A caching strategy relying solely on local variables for every input and output.

Implements the CacheStrategy interface.

GraphComputing.Machine — Type

Machine

A representation of a machine to execute on. Contains information about its architecture (CPUs, GPUs, maybe more). This representation can be used to make a more accurate cost prediction of a DAG state.

See also: Scheduler

GraphComputing.CACHE_STRATEGIES — Constant

CACHE_STRATEGIES::Dict{Type{AbstractDevice}, Symbol}

Global dictionary of available caching strategies per device. Each implementation of AbstractDevice should add its available strategies to the dictionary.

See also: strategies

GraphComputing.DEVICE_TYPES — Constant

DEVICE_TYPES::Vector{Type}

Global vector of available and implemented device types. Each implementation of a AbstractDevice should add its concrete type to this vector.

See also: device_types, get_devices

GraphComputing.default_strategy — Function

default_strategy(deviceType::Type{T}) where {T <: AbstractDevice}

Interface function that must be implemented for every subtype of AbstractDevice. Returns the default CacheStrategy to use on the given device type. See also: cache_strategy, set_cache_strategy

GraphComputing.gen_access_expr — Function

gen_access_expr(device::AbstractDevice, symbol::Symbol)

Interface function that must be implemented for every subtype of AbstractDevice and at least one CacheStrategy. Return an Expr or QuoteNode accessing the variable identified by [symbol].

GraphComputing.gen_cache_init_code — Function

gen_cache_init_code(device::AbstractDevice)

Interface function that must be implemented for every subtype of AbstractDevice and at least one CacheStrategy. Returns an Expr initializing this device's variable cache.

The strategy is a symbol

GraphComputing.get_devices — Function

get_devices(t::Type{T}; verbose::Bool) where {T <: AbstractDevice}

Interface function that must be implemented for every subtype of AbstractDevice. Returns a Vector{Type} of the devices for the given AbstractDevice Type available on the current machine.

GraphComputing.measure_device! — Function

measure_device!(device::AbstractDevice; verbose::Bool)

Interface function that must be implemented for every subtype of AbstractDevice. Measures the compute speed of the given device and writes into it.

Detect

GraphComputing.get_machine_info — Method

get_machine_info(verbose::Bool)

Return the Machine currently running on. The parameter verbose defaults to true when interactive.

Measure

GraphComputing.measure_devices! — Method

measure_devices(machine::Machine; verbose::Bool)

Measure FLOPS, RAM, cache sizes and what other properties can be extracted for the devices in the given machine.

GraphComputing.measure_transfer_rates! — Method

measure_transfer_rates(machine::Machine; verbose::Bool)

Measure the transfer rates between devices in the machine.

Implementations

General

GraphComputing.cache_strategy — Method

cache_strategy(device::AbstractDevice)

Returns the cache strategy set for this device.

GraphComputing.cpu_st — Method

cpu_st()

A function returning a Machine that only has a single thread of one CPU. It is the simplest machine definition possible and produces a simple function when used with get_compute_function.

GraphComputing.device_types — Method

device_types()

Return a vector of available and implemented device types.

See also: DEVICE_TYPES

GraphComputing.entry_device — Method

entry_device(machine::Machine)

Return the "entry" device, i.e., the device that starts CPU threads and GPU kernels, and takes input values and returns the output value.

GraphComputing.set_cache_strategy — Method

set_cache_strategy(device::AbstractDevice, cacheStrategy::CacheStrategy)

Sets the device's cache strategy. After this call, cache_strategy should return cacheStrategy on the given device.

GraphComputing.strategies — Method

strategies(t::Type{T}) where {T <: AbstractDevice}

Return a vector of available CacheStrategys for the given AbstractDevice. The caching strategies are used in code generation.

NUMA

GraphComputing.NumaNode — Type

NumaNode <: AbstractCPU

Representation of a specific CPU that code can run on. Implements the AbstractDevice interface.

GraphComputing._gen_access_expr — Method

_gen_access_expr(device::NumaNode, ::Dictionary, symbol::Symbol)

Internal function for dispatch, used in gen_access_expr.

GraphComputing._gen_access_expr — Method

_gen_access_expr(device::NumaNode, ::LocalVariables, symbol::Symbol)

Internal function for dispatch, used in gen_access_expr.

GraphComputing.gen_access_expr — Method

gen_access_expr(device::NumaNode, symbol::Symbol)

Generate code to access the variable designated by symbol on a NumaNode, using the CacheStrategy set in the device.

GraphComputing.gen_cache_init_code — Method

gen_cache_init_code(device::NumaNode)

Generate code for initializing the LocalVariables strategy on a NumaNode.

GraphComputing.get_devices — Method

get_devices(deviceType::Type{T}; verbose::Bool) where {T <: NumaNode}

Return a Vector of NumaNodes available on the current machine. If verbose is true, print some additional information.

CUDA

GraphComputing.CUDAGPU — Type

CUDAGPU <: AbstractGPU

Representation of a specific CUDA GPU that code can run on. Implements the AbstractDevice interface.

GraphComputing.get_devices — Method

get_devices(deviceType::Type{T}; verbose::Bool) where {T <: CUDAGPU}

Return a Vector of CUDAGPUs available on the current machine. If verbose is true, print some additional information.

ROCm

GraphComputing.ROCmGPU — Type

ROCmGPU <: AbstractGPU

Representation of a specific AMD GPU that code can run on. Implements the AbstractDevice interface.

GraphComputing.get_devices — Method

get_devices(deviceType::Type{T}; verbose::Bool = false) where {T <: ROCmGPU}

Return a Vector of ROCmGPUs available on the current machine. If verbose is true, print some additional information.

oneAPI